Fix issue with FFI boolean parameters being the wrong size for dart (#261)
* Fix issue with FFI boolean parameters being the wrong size for dart

* fix moved node download url

+++ Sources/FishyJoesIotaRuntime/IotaConverter.swift
@@ -77,8 +77,7 @@ extension VoidConverter: IotaConverter {
 }
 
 extension Bool: IotaConverter {
-    // TODO: possible to marshall as 1-byte, but requires weird annotations
-    public typealias CType = UInt32
+    public typealias CType = UInt8
 
     public typealias ValueMethod = @convention(c) (foreignObject, foreignOutExn) -> CType
